#77 modifying wrapper
